Update deliverable_2.py
(Posting another commit, with more comments, since I forgot to add a description to last one).

I turned the data into arrays, and filtered it to only keep the study hours above 2 hours, and its respective exam score. In order to do that, I used condition and "for loops" to go through the arrays to find the positions where the study hours were larger than 2.